requirements.txt
psycopg2-binary
python-dotenv
opencv-python
face_recognition
numpy